+++ FollowerCommands/MateFollower.cs
@@ -14,14 +14,16 @@ using UnityEngine;
 
 namespace MatingTentMod.FollowerCommands;
 
-public class MateCommand : CustomFollowerCommand
+public class MateFollower : CustomFollowerCommand
 {
-    public override string InternalName => "MATE_COMMAND";
-    public override Sprite CommandIcon => TextureHelper.CreateSpriteFromPath(Path.Combine(Plugin.PluginPath, "Assets/MatingTent.png"));
+    public override string InternalName => "MATE_COMMAND_FOLLOWER";
+
+    public override Sprite CommandIcon =>
+        TextureHelper.CreateSpriteFromPath(Path.Combine(Plugin.PluginPath, "Assets/MatingTent.png"));
 
     public override string GetTitle(Follower follower)
     {
-        return "Mate";
+        return "Mate with Follower";
     }
 
     public override string GetDescription(Follower follower)
@@ -31,14 +33,22 @@ public class MateCommand : CustomFollowerCommand
 
     public override bool ShouldAppearFor(Follower follower)
     {
-        if (follower.Brain.Info.OldAge || follower.Brain.Info.Age < 18) return false;
-        return Structure.Structures.FirstOrDefault<Structure>(str => str.Structure_Info != null && str.Structure_Info.Type == CustomStructureManager.GetStructureByType<MatingTent>()) != null;
+        if (follower.Brain.Info.OldAge || follower.Brain.Info.Age < 18)
+        {
+            return false;
+        }
+
+        return Structure.Structures.FirstOrDefault(str =>
+            str.Structure_Info != null &&
+            str.Structure_Info.Type == CustomStructureManager.GetStructureByType<MatingTent>()) != null;
     }
 
     public override bool IsAvailable(Follower follower)
     {
         return TaskUtils.GetAvailableStructureOfType<MatingTent>() != null && FollowerBrain
-            .AllAvailableFollowerBrains().Any(x => x != follower.Brain && !FollowerManager.FollowerLocked(x.Info.ID, true) && x.CurrentTask is not MateLead && x.CurrentTask is not MateFollow);
+            .AllAvailableFollowerBrains().Any(x =>
+                x != follower.Brain && !FollowerManager.FollowerLocked(x.Info.ID, true) &&
+                x.CurrentTask is not MateLead && x.CurrentTask is not MateFollow);
     }
 
     public override string GetLockedDescription(Follower follower)
@@ -71,7 +81,7 @@ public class MateCommand : CustomFollowerCommand
             x != interaction.follower.Brain && !FollowerManager.FollowerLocked(x.Info.ID, true) &&
             x.Info.CursedState != CustomCursedStates.CHILD && x.Info.CursedState != Thought.OldAge).ToList();
         mateFollowerSelect.Show(validBrains);
-        mateFollowerSelect.OnFollowerSelected = (selectedFollower) =>
+        mateFollowerSelect.OnFollowerSelected = selectedFollower =>
         {
             if (selectedFollower == null || selectedFollower.ID == interaction.follower.Brain.Info.ID)
             {
@@ -85,10 +95,8 @@ public class MateCommand : CustomFollowerCommand
                 FollowerBrain selectedFollowerBrain = FollowerBrain.FindBrainByID(selectedFollower.ID);
                 MateActions.BeginMate(interaction.follower.Brain, selectedFollowerBrain);
                 interaction.Close(true, reshowMenu: false);
-                interaction.Close(true, reshowMenu: false);
             }));
         };
         mateFollowerSelect.OnCancel = delegate { interaction.Close(true, reshowMenu: false); };
     }
-
 }