+++ Plugin.cs
@@ -7,49 +7,50 @@ using MatingTentMod.FollowerCommands;
 using MatingTentMod.Structures;
 using System.IO;
 
-namespace MatingTentMod
+namespace MatingTentMod;
+
+[BepInPlugin(PluginGuid, PluginName, PluginVer)]
+[BepInDependency("io.github.xhayper.COTL_API")]
+[HarmonyPatch]
+public class Plugin : BaseUnityPlugin
 {
-    [BepInPlugin(PluginGuid, PluginName, PluginVer)]
-    [BepInDependency("io.github.xhayper.COTL_API")]
-    [HarmonyPatch]
-    public class Plugin : BaseUnityPlugin
+    public const string PluginGuid = "IngoH.cotl.MatingTentMod";
+    public const string PluginName = "MatingTentMod";
+    public const string PluginVer = "1.0.0";
+
+    internal static ManualLogSource Log;
+    internal static readonly Harmony Harmony = new(PluginGuid);
+
+    internal static string PluginPath;
+
+    internal static MatingTent MatingTent;
+    internal static MateFollower MateCommand;
+    internal static MateLamb MateLamb;
+
+    private void Awake()
     {
-        public const string PluginGuid = "IngoH.cotl.MatingTentMod";
-        public const string PluginName = "MatingTentMod";
-        public const string PluginVer = "0.5.1";
+        this.Logger.LogInfo($"Loaded {PluginName}!");
+        Log = this.Logger;
 
-        internal static ManualLogSource Log;
-        internal readonly static Harmony Harmony = new(PluginGuid);
+        PluginPath = Path.GetDirectoryName(this.Info.Location);
 
-        internal static string PluginPath;
+        MatingTent = new MatingTent();
+        CustomStructureManager.Add(new MatingTent());
+        MateCommand = new MateFollower();
+        CustomFollowerCommandManager.Add(MateCommand);
+        MateLamb = new MateLamb();
+        CustomFollowerCommandManager.Add(MateLamb);
+    }
 
-        internal static MatingTent MatingTent;
-        internal static MateCommand MateCommand;
+    private void OnEnable()
+    {
+        Harmony.PatchAll();
+        this.Logger.LogInfo($"Loaded {PluginName}!");
+    }
 
-        private void Awake()
-        {
-            Logger.LogInfo($"Loaded {PluginName}!");
-            Plugin.Log = base.Logger;
-
-            PluginPath = Path.GetDirectoryName(Info.Location);
-
-            MatingTent = new MatingTent();
-            CustomStructureManager.Add(new MatingTent());
-            MateCommand = new MateCommand();
-            CustomFollowerCommandManager.Add(MateCommand);
-
-        }
-
-        private void OnEnable()
-        {
-            Harmony.PatchAll();
-            Logger.LogInfo($"Loaded {PluginName}!");
-        }
-
-        private void OnDisable()
-        {
-            Harmony.UnpatchSelf();
-            Logger.LogInfo($"Unloaded {PluginName}!");
-        }
+    private void OnDisable()
+    {
+        Harmony.UnpatchSelf();
+        this.Logger.LogInfo($"Unloaded {PluginName}!");
     }
 }